requirements.txt
langchain
langchain-community
tiktoken
faiss-cpu
streamlit
python-dotenv
pypdf
docx2txt
langchain-google-genai
